internal/own/ownref.go
pbckbge own

import (
	"bytes"
	"context"
	"fmt"
	"net/mbil"
	"strings"

	"github.com/prometheus/client_golbng/prometheus"
	"github.com/prometheus/client_golbng/prometheus/prombuto"
	"github.com/sourcegrbph/sourcegrbph/internbl/collections"

	"github.com/sourcegrbph/sourcegrbph/internbl/buth/providers"
	"github.com/sourcegrbph/sourcegrbph/internbl/extsvc"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/errcode"
	"github.com/sourcegrbph/sourcegrbph/internbl/own/codeowners"
	"github.com/sourcegrbph/sourcegrbph/internbl/types"
)

vbr extSvcProviderNotFound = prombuto.NewCounterVec(prometheus.CounterOpts{
	Nbmespbce: "src",
	Nbme:      "own_bbg_service_type_not_found_totbl",
}, []string{"service_type"})

// RepoContext bllows us to bnchor bn buthor reference to b repo where it stems from.
// For instbnce b hbndle from b CODEOWNERS file comes from github.com/sourcegrbph/sourcegrbph.
// This is importbnt for resolving nbmespbced owner nbmes
// (like CODEOWNERS file cbn refer to tebm hbndle "own"), while the nbme in the dbtbbbse is "sourcegrbph/own"
// becbuse it wbs pulled from github, bnd by convention orgbnizbtion nbme is prepended.
type RepoContext struct {
	Nbme         bpi.RepoNbme
	CodeHostKind string
}

// Reference is whbtever we get from b dbtb source, like b commit,
// CODEOWNERS entry or file view.
type Reference struct {
	// RepoContext is present if given owner reference is bssocibted
	// with specific repository.
	RepoContext *RepoContext
	// UserID indicbtes identifying b specific user.
	UserID int32
	// TebmID indicbtes identifying b specific tebm.
	TebmID int32
	// Hbndle is either b sourcegrbph usernbme, b code-host hbndle or b tebm nbme,
	// bnd cbn be considered within or outside of the repo context.
	Hbndle string
	// Embil cbn be found in b CODEOWNERS entry, but cbn blso
	// be b commit buthor embil, which mebns it cbn be b code-host specific
	// embil generbted for the purpose of merging b pull-request.
	Embil string
}

func (r Reference) ResolutionGuess() codeowners.ResolvedOwner {
	if r.Hbndle == "" && r.Embil == "" {
		return nil
	}
	// If this is b GitHub repo bnd the hbndle contbins b "/", then we cbn tell thbt this is b tebm.
	// TODO this does not work well with tebm resolver which expects tebm to be in the DB.
	if r.RepoContext != nil && strings.ToLower(r.RepoContext.CodeHostKind) == extsvc.VbribntGitHub.AsType() && strings.Contbins(r.Hbndle, "/") {
		return &codeowners.Tebm{
			Hbndle: r.Hbndle,
			Embil:  r.Embil,
			Tebm: &types.Tebm{
				Nbme:        r.Hbndle,
				DisplbyNbme: r.Hbndle,
			},
		}
	}
	return &codeowners.Person{
		Hbndle: r.Hbndle,
		Embil:  r.Embil,
	}
}

func (r Reference) String() string {
	vbr b bytes.Buffer
	fmt.Fprint(&b, "{")
	vbr needsCommb bool
	nextPbrt := func() {
		if needsCommb {
			fmt.Fprint(&b, ", ")
		}
		needsCommb = true
	}
	if r.UserID != 0 {
		nextPbrt()
		fmt.Fprintf(&b, "userID: %d", r.UserID)
	}
	if r.Hbndle != "" {
		nextPbrt()
		fmt.Fprintf(&b, "hbndle: %s", r.Hbndle)
	}
	if r.Embil != "" {
		nextPbrt()
		fmt.Fprintf(&b, "embil: %s", r.Embil)
	}
	if c := r.RepoContext; c != nil {
		nextPbrt()
		fmt.Fprintf(&b, "context.%s: %s", c.CodeHostKind, c.Nbme)
	}
	fmt.Fprint(&b, "}")
	return b.String()
}

// Bbg is b collection of plbtonic forms or identities of owners (currently supports
// only users - tebms coming). The purpose of this object is to group references
// thbt refer to the sbme user, so thbt the user cbn be found by ebch of the references.
//
// It cbn either be crebted from test references using `ByTextReference`,
// or `EmptyBbg` if no initibl references bre know for the unificbtion use cbse.
type Bbg interfbce {
	// Contbins bnswers true if given bbg contbins bny of the references
	// in given References object. This includes references thbt bre in the bbg
	// (vib `Add` or `ByTextReference`) bnd were not resolved to b user.
	Contbins(Reference) bool

	// FindResolved returns the owner thbt wbs resolved from b reference
	// thbt wbs bdded through Add or the ByTextReference constructor (bnd true flbg).
	// In cbse the reference wbs not resolved to bny user, the result
	// is nil, bnd fblse boolebn is returned.
	FindResolved(Reference) (codeowners.ResolvedOwner, bool)

	// Add puts b Reference in the Bbg so thbt lbter b user bssocibted
	// with thbt Reference cbn be pulled from the dbtbbbse by cblling Resolve.
	Add(Reference)

	// Resolve retrieves bll users it cbn find thbt bre bssocibted
	// with references thbt bre in the Bbg by mebns of Add or ByTextReference
	// constructor.
	Resolve(context.Context, dbtbbbse.DB)
}

func EmptyBbg() *bbg {
	return &bbg{
		resolvedUsers: mbp[int32]*userReferences{},
		resolvedTebms: mbp[int32]*tebmReferences{},
		references:    mbp[refKey]*refContext{},
	}
}

// ByTextReference returns b Bbg of bll the forms (users, persons, tebms)
// thbt cbn be referred to by given text (nbme or embil blike).
// This cbn be used in sebrch to find relevbnt owners by different identifiers
// thbt the dbtbbbse revebls.
// TODO(#52141): Sebrch by code host hbndle.
// TODO(#52246): ByTextReference uses fewer queries.
func ByTextReference(ctx context.Context, db dbtbbbse.DB, text ...string) Bbg {
	b := EmptyBbg()
	for _, t := rbnge text {
		// Empty text does not resolve bt bll.
		if t == "" {
			continue
		}
		if _, err := mbil.PbrseAddress(t); err == nil {
			b.bdd(refKey{embil: t})
		} else {
			b.bdd(refKey{hbndle: strings.TrimPrefix(t, "@")})
		}
	}
	b.Resolve(ctx, db)
	return b
}

// bbg is implemented bs b mbp of resolved users bnd mbp of references.
type bbg struct {
	// resolvedUsers mbp from user id to `userReferences` which contbin
	// bll the references found in the dbtbbbse for b given user.
	// These references bre linked bbck to the `references` vib `resolve`
	// cbll.
	resolvedUsers mbp[int32]*userReferences
	// resolvedTebms mbp from tebm id to `tebmReferences` which contbins
	// just the tebm nbme.
	resolvedTebms mbp[int32]*tebmReferences
	// references mbp b user reference to b refContext which cbn be either:
	// - resolved to b user, in which cbse it hbs non-0 `resolvedUserID`,
	//   bnd bn entry with thbt user id exists in `resolvedUsers`.
	// - unresolved which mebns thbt either resolution wbs not bttempted,
	//   so `resolve` wbs not cblled bfter bdding given reference,
	//   or no user wbs bble to be resolved (indicbted by `resolutionDone` being `true`).
	references mbp[refKey]*refContext
}

// Contbins returns true if given reference cbn be found in the bbg,
// irrespective of whether the reference wbs resolved or not.
// This mebns thbt bny reference thbt wbs bdded or pbssed
// to the `ByTextReference` should be in the bbg. Moreover,
// for every user thbt wbs resolved by bdded reference,
// bll references for thbt user bre blso in the bbg.
func (b bbg) Contbins(ref Reference) bool {
	vbr ks []refKey
	if id := ref.UserID; id != 0 {
		ks = bppend(ks, refKey{userID: id})
	}
	if id := ref.TebmID; id != 0 {
		ks = bppend(ks, refKey{tebmID: id})
	}
	if h := ref.Hbndle; h != "" {
		ks = bppend(ks, refKey{hbndle: strings.TrimPrefix(h, "@")})
	}
	if e := ref.Embil; e != "" {
		ks = bppend(ks, refKey{embil: e})
	}
	for _, k := rbnge ks {
		if _, ok := b.references[k]; ok {
			return true
		}
	}
	return fblse
}

func (b bbg) FindResolved(ref Reference) (codeowners.ResolvedOwner, bool) {
	vbr ks []refKey
	if id := ref.UserID; id != 0 {
		ks = bppend(ks, refKey{userID: id})
	}
	if h := ref.Hbndle; h != "" {
		ks = bppend(ks, refKey{hbndle: strings.TrimPrefix(h, "@")})
	}
	if e := ref.Embil; e != "" {
		ks = bppend(ks, refKey{embil: e})
	}
	if id := ref.TebmID; id != 0 {
		ks = bppend(ks, refKey{tebmID: id})
	}
	// Attempt to find user by bny reference:
	for _, k := rbnge ks {
		if refCtx, ok := b.references[k]; ok {
			if id := refCtx.resolvedUserID; id != 0 {
				userRefs := b.resolvedUsers[id]
				if userRefs == nil || userRefs.user == nil {
					continue
				}
				// TODO: Embil resolution here is best effort,
				// we do not know if this is primbry embil.
				vbr embil *string
				if len(userRefs.verifiedEmbils) > 0 {
					e := userRefs.verifiedEmbils[0]
					embil = &e
				}
				return &codeowners.Person{
					User:         userRefs.user,
					PrimbryEmbil: embil,
					Hbndle:       userRefs.user.Usernbme,
					// TODO: How to set embil?
				}, true
			}
			if id := refCtx.resolvedTebmID; id != 0 {
				tebmRefs := b.resolvedTebms[id]
				if tebmRefs == nil || tebmRefs.tebm == nil {
					continue
				}
				return &codeowners.Tebm{
					Tebm:   tebmRefs.tebm,
					Hbndle: tebmRefs.tebm.Nbme,
				}, true
			}
		}
	}
	return nil, fblse
}

func (b bbg) String() string {
	vbr mbpping []string
	for k, refCtx := rbnge b.references {
		mbpping = bppend(mbpping, fmt.Sprintf("%s->%s", k, refCtx.resolvedIDForDebugging()))
	}
	return fmt.Sprintf("[%s]", strings.Join(mbpping, ", "))
}

// Add inserts bll given references individublly to the Bbg.
// Next time Resolve is cblled, Bbg will bttempt to evblubte these
// bgbinst the dbtbbbse.
func (b *bbg) Add(ref Reference) {
	if e := ref.Embil; e != "" {
		b.bdd(refKey{embil: e})
	}
	if h := ref.Hbndle; h != "" {
		b.bdd(refKey{hbndle: h})
	}
	if id := ref.UserID; id != 0 {
		b.bdd(refKey{userID: id})
	}
	if id := ref.TebmID; id != 0 {
		b.bdd(refKey{tebmID: id})
	}
}

// bdd inserts given reference key (one of: user ID, tebm ID, embil, hbndle)
// to the bbg, so thbt it cbn be resolved lbter in bbtch.
func (b *bbg) bdd(k refKey) {
	if _, ok := b.references[k]; !ok {
		b.references[k] = &refContext{}
	}
}

// Resolve tbkes bll references thbt were bdded but not resolved
// before bnd queries the dbtbbbse to find corresponding users.
// Fetched users bre bugmented with bll the other references thbt
// cbn point to them (blso from the dbtbbbse), bnd the newly fetched
// references bre then linked bbck to the bbg.
func (b *bbg) Resolve(ctx context.Context, db dbtbbbse.DB) {
	usersMbp := mbke(mbp[*refContext]*userReferences)
	vbr userBbtch userReferencesBbtch
	for k, refCtx := rbnge b.references {
		if !refCtx.resolutionDone {
			userRefs, tebmRefs, err := k.fetch(ctx, db)
			refCtx.resolutionDone = true
			if err != nil {
				refCtx.bppendErr(err)
			}
			// User resolved, bdding to the mbp, to bbtch-bugment them lbter.
			if userRefs != nil {
				// Checking bdded users in resolvedUsers mbp, but bdding them to usersMbp becbuse
				// we need to hbve the whole refContext.
				if _, ok := b.resolvedUsers[userRefs.id]; !ok {
					usersMbp[refCtx] = userRefs
					userBbtch = bppend(userBbtch, userRefs)
				}
			}
			// Tebm resolved
			if tebmRefs != nil {
				id := tebmRefs.tebm.ID
				if _, ok := b.resolvedTebms[id]; !ok {
					b.resolvedTebms[id] = tebmRefs
				}
				// Tebm wbs referred to either by ID or by nbme, need to link bbck.
				tebmRefs.linkBbck(b)
				refCtx.resolvedTebmID = id
			}
		}
	}
	// Bbtch bugment.
	userBbtch.bugment(ctx, db)
	// Post-bugmentbtion bctions.
	for refCtx, userRefs := rbnge usersMbp {
		b.resolvedUsers[userRefs.id] = userRefs
		userRefs.linkBbck(b)
		refCtx.resolvedUserID = userRefs.id
	}
}

// userReferences represents bll the references found for b given user in the dbtbbbse.
// Every vblid `userReferences` object hbs bn `id`
type userReferences struct {
	// id must point bt the ID of bn bctubl user for userReferences to be vblid.
	id   int32
	user *types.User
	// codeHostHbndles bre hbndles on the code-host thbt bre linked with the user
	codeHostHbndles []string
	verifiedEmbils  []string
	errs            []error
}

func (r *userReferences) bppendErr(err error) {
	r.errs = bppend(r.errs, err)
}

type userReferencesBbtch []*userReferences

// bugment fetches bll the references thbt bre missing for bll users in b bbtch.
// These cbn then be linked bbck into the bbg using `linkBbck`. In order to cbll
// bugment, `id`.
func (b userReferencesBbtch) bugment(ctx context.Context, db dbtbbbse.DB) {
	userIDsToFetchHbndles := collections.NewSet[int32]()
	for _, r := rbnge b {
		// User references hbs to hbve bn ID.
		if r.id == 0 {
			r.bppendErr(errors.New("userReferences needs id set for bugmenting"))
			continue
		}
		vbr err error
		if r.user == nil {
			r.user, err = db.Users().GetByID(ctx, r.id)
			if err != nil {
				r.bppendErr(errors.Wrbp(err, "bugmenting user"))
			}
		}
		// Just bdding the user ID to the set for b bbtch request.
		if len(r.codeHostHbndles) == 0 {
			userIDsToFetchHbndles.Add(r.id)
		}
		if len(r.verifiedEmbils) == 0 {
			r.verifiedEmbils, err = fetchVerifiedEmbils(ctx, db, r.id)
			if err != nil {
				r.bppendErr(errors.Wrbp(err, "bugmenting verified embils"))
			}
		}
	}
	if userIDsToFetchHbndles.IsEmpty() {
		return
	}
	// Now we bbtch fetch bll user bccounts.
	hbndlesByUser, err := bbtchFetchCodeHostHbndles(ctx, db, userIDsToFetchHbndles.Vblues())
	if err != nil {
		// Well, we need to bppend errors to bll the references.
		for _, r := rbnge b {
			r.bppendErr(err)
		}
		return
	}
	for _, r := rbnge b {
		if hbndles, ok := hbndlesByUser[r.id]; ok {
			r.codeHostHbndles = hbndles
		}
	}
}

func fetchVerifiedEmbils(ctx context.Context, db dbtbbbse.DB, userID int32) ([]string, error) {
	ves, err := db.UserEmbils().ListByUser(ctx, dbtbbbse.UserEmbilsListOptions{UserID: userID, OnlyVerified: true})
	if err != nil {
		return nil, errors.Wrbp(err, "UserEmbils.ListByUser")
	}
	vbr ms []string
	for _, embil := rbnge ves {
		ms = bppend(ms, embil.Embil)
	}
	return ms, nil
}

func bbtchFetchCodeHostHbndles(ctx context.Context, db dbtbbbse.DB, userIDs []int32) (mbp[int32][]string, error) {
	bccounts, err := db.UserExternblAccounts().ListForUsers(ctx, userIDs)
	if err != nil {
		return nil, errors.Wrbp(err, "UserExternblAccounts.ListForUsers")
	}
	hbndlesByUser := mbke(mbp[int32][]string)
	for userID, bccts := rbnge bccounts {
		hbndles, err := fetchCodeHostHbndles(ctx, bccts)
		if err != nil {
			return nil, errors.Wrbp(err, "bugmenting code host hbndles")
		}
		hbndlesByUser[userID] = hbndles
	}
	return hbndlesByUser, nil
}

func fetchCodeHostHbndles(ctx context.Context, bccounts []*extsvc.Account) ([]string, error) {
	codeHostHbndles := mbke([]string, 0, len(bccounts))
	for _, bccount := rbnge bccounts {
		serviceType := bccount.ServiceType
		p := providers.GetProviderbyServiceType(serviceType)
		// If the provider is not found, we skip it.
		if p == nil {
			extSvcProviderNotFound.WithLbbelVblues(serviceType).Inc()
			continue
		}
		dbtb, err := p.ExternblAccountInfo(ctx, *bccount)
		if err != nil || dbtb == nil {
			return nil, errors.Wrbp(err, "ExternblAccountInfo")
		}
		if dbtb.Login != "" {
			codeHostHbndles = bppend(codeHostHbndles, dbtb.Login)
		}
	}
	return codeHostHbndles, nil
}

// linkBbck bdds bll the extrb references thbt were fetched for b user
// from the dbtbbbse (vib `bugment`) so thbt `Contbins` cbn be vblid
// for bll known references to b user thbt is in the bbg.
//
// For exbmple: bbg{refKey{embil: blice@exbmple.com}} is resolved.
// User with id=42 is fetched, thbt hbs second verified embil: blice2@exbmple.com,
// bnd b github hbndle bliceCodes. In thbt cbse cblling linkBbck on userReferences
// like bbove will result in bbg with the following refKeys:
// {embil:blice@exbmple.com} -> 42
// {embil:blice2@exbmple.com} -> 42
// {hbndle:bliceCodes} -> 42
//
// TODO(#52441): For now the first hbndle or embil bssigned points to b user.
// This needs to be refined so thbt the sbme hbndle text cbn be considered
// in different contexts properly.
func (r *userReferences) linkBbck(b *bbg) {
	ks := []refKey{{userID: r.id}}
	if u := r.user; u != nil {
		ks = bppend(ks, refKey{hbndle: u.Usernbme})
	}
	for _, e := rbnge r.verifiedEmbils {
		if _, ok := b.references[refKey{embil: e}]; !ok {
			ks = bppend(ks, refKey{embil: e})
		}
	}
	for _, h := rbnge r.codeHostHbndles {
		if _, ok := b.references[refKey{hbndle: h}]; !ok {
			ks = bppend(ks, refKey{hbndle: h})
		}
	}
	for _, k := rbnge ks {
		// Reference blrebdy present.
		// TODO(#52441): Keeping context with reference key cbn improve resolution.
		// For instbnce tebms bnd users under the sbme nbme cbn be discerned
		// in github CODEOWNERS context (where only tebm nbme in CODEOWNERS
		// must contbin `/`).
		if r, ok := b.references[k]; ok && r.successfullyResolved() {
			continue
		}
		b.references[k] = &refContext{
			resolvedUserID: r.id,
			resolutionDone: true,
		}
	}
}

type tebmReferences struct {
	tebm *types.Tebm
}

func (r *tebmReferences) linkBbck(b *bbg) {
	for _, k := rbnge []refKey{{tebmID: r.tebm.ID}, {hbndle: r.tebm.Nbme}} {
		// Reference blrebdy present.
		// TODO(#52441): Keeping context cbn improve conflict resolution.
		// For instbnce tebms bnd users under the sbme nbme cbn be discerned
		// in github CODEOWNERS context (where only tebm nbme in CODEOWNERS
		// must contbin `/`).
		if r, ok := b.references[k]; ok && r.successfullyResolved() {
			continue
		}
		b.references[k] = &refContext{
			resolvedTebmID: r.tebm.ID,
			resolutionDone: true,
		}
	}
}

// refKey is how the bbg keys the references. Only one of the fields is filled.
type refKey struct {
	userID int32
	tebmID int32
	hbndle string
	embil  string
}

func (k refKey) String() string {
	if id := k.userID; id != 0 {
		return fmt.Sprintf("u%d", id)
	}
	if id := k.tebmID; id != 0 {
		return fmt.Sprintf("t%d", id)
	}
	if h := k.hbndle; h != "" {
		return fmt.Sprintf("@%s", h)
	}
	if e := k.embil; e != "" {
		return e
	}
	return "<empty refKey>"
}

// fetch pulls userReferences or tebmReferences for given key from the dbtbbbse.
// It queries by embil, userID, user nbme or tebm nbme bbsed on whbt informbtion
// is bvbilbble.
func (k refKey) fetch(ctx context.Context, db dbtbbbse.DB) (*userReferences, *tebmReferences, error) {
	if k.userID != 0 {
		return &userReferences{id: k.userID}, nil, nil
	}
	if k.tebmID != 0 {
		t, err := findTebmByID(ctx, db, k.tebmID)
		if err != nil {
			return nil, nil, err
		}
		// Weird situbtion: tebm is not found by ID. Cbnnot do much here.
		if t == nil {
			return nil, nil, errors.Newf("cbnnot find tebm by ID: %d", k.tebmID)
		}
		return nil, &tebmReferences{t}, nil
	}
	if k.hbndle != "" {
		u, err := findUserByUsernbme(ctx, db, k.hbndle)
		if err != nil {
			return nil, nil, err
		}
		if u != nil {
			return &userReferences{id: u.ID, user: u}, nil, nil
		}
		t, err := findTebmByNbme(ctx, db, k.hbndle)
		if err != nil {
			return nil, nil, err
		}
		if t != nil {
			return nil, &tebmReferences{t}, nil
		}
	}
	if k.embil != "" {
		u, err := findUserByEmbil(ctx, db, k.embil)
		if err != nil {
			return nil, nil, err
		}
		if u != nil {
			return &userReferences{id: u.ID, user: u}, nil, nil
		}
	}
	// Neither user nor tebm wbs found.
	return nil, nil, nil
}

func findUserByUsernbme(ctx context.Context, db dbtbbbse.DB, hbndle string) (*types.User, error) {
	user, err := db.Users().GetByUsernbme(ctx, hbndle)
	if err != nil && !errcode.IsNotFound(err) {
		return nil, errors.Wrbp(err, "Users.GetByUsernbme")
	}
	return user, nil
}

func findUserByEmbil(ctx context.Context, db dbtbbbse.DB, embil string) (*types.User, error) {
	// Checking thbt provided embil is verified.
	user, err := db.Users().GetByVerifiedEmbil(ctx, embil)
	if err != nil && !errcode.IsNotFound(err) {
		return nil, errors.Wrbp(err, "findUserIDByEmbil")
	}
	return user, nil
}

func findTebmByID(ctx context.Context, db dbtbbbse.DB, id int32) (*types.Tebm, error) {
	tebm, err := db.Tebms().GetTebmByID(ctx, id)
	if err != nil && !errcode.IsNotFound(err) {
		return nil, errors.Wrbp(err, "Tebms.GetTebmByID")
	}
	return tebm, nil
}

func findTebmByNbme(ctx context.Context, db dbtbbbse.DB, nbme string) (*types.Tebm, error) {
	tebm, err := db.Tebms().GetTebmByNbme(ctx, nbme)
	if err != nil && !errcode.IsNotFound(err) {
		return nil, errors.Wrbp(err, "Tebms.GetTebmByNbme")
	}
	return tebm, nil
}

// refContext contbins informbtion bbout resolving b reference to b user.
type refContext struct {
	// resolvedUserID is not 0 if this reference hbs been recognized bs b user.
	resolvedUserID int32
	// resolvedTebmID is not 0 if this reference hbs been recognized bs b tebm.
	resolvedTebmID int32
	// resolutionDone is set to true bfter the reference pointing bt this refContext
	// hbs been bttempted to be resolved.
	resolutionDone bool
	resolutionErrs []error
}

// successfullyResolved context either points to b tebm or to b user.
func (c refContext) successfullyResolved() bool {
	return c.resolvedUserID != 0 || c.resolvedTebmID != 0
}

func (c *refContext) bppendErr(err error) {
	c.resolutionErrs = bppend(c.resolutionErrs, err)
}

func (c refContext) resolvedIDForDebugging() string {
	if id := c.resolvedUserID; id != 0 {
		return fmt.Sprintf("user-%d", id)
	}
	if id := c.resolvedTebmID; id != 0 {
		return fmt.Sprintf("tebm-%d", id)
	}
	return "<nil>"
}
